# docs/swagger.yaml
definitions:
  base.PageResp:
    properties:
      currentPage:
        description: 当前第几页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 分页大小
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
  base.ReqId:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  base.ReqIds:
    properties:
      ids:
        description: 多id
        items:
          type: integer
        type: array
    type: object
  base.Resp:
    properties:
      code:
        description: 返回码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
      reqId:
        description: '`json:"请求id"`'
        type: string
    type: object
...
  dto.SysApiDto:
    properties:
      id:
        description: 主键
        type: integer
      method:
        type: string
      path:
        type: string
      permType:
        type: integer
      status:
        type: integer
      title:
        type: string
    type: object
  dto.SysApiGetPageReq:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      status:
        type: integer
    type: object
  dto.SysMenuApiDto:
    properties:
      sysApiId:
        type: integer
      sysMenuId:
        type: integer
    type: object
  dto.SysMenuApiGetPageReq:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  models.SysApi:
    properties:
      id:
        description: 主键
        type: integer
      method:
        description: 请求类型
        type: string
      path:
        description: 请求地址
        type: string
      permType:
        description: 权限类型（1：无需认证 2:须token 3：须鉴权）
        type: integer
      status:
        description: 状态 3 DEF 2 OK 1 del
        type: integer
      title:
        description: 标题
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  models.SysMenuApi:
    properties:
      sysApiId:
        type: integer
      sysMenuId:
        type: integer
    type: object
...
info:
  contact:
    email: lisite199505@gmail.com
    name: lei
    url: https://github.com/leijeng/huo-admin
  description: 致力于做一个开发快速，运行稳定的框架
  title: Huo API
  version: V1.0.1
paths:
  /v2/admin/sys/sys-api/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysApiDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysApi'
              type: object
      security:
      - Bearer: []
      summary: 创建SysApi
      tags:
      - sys-SysApi
  /v2/admin/sys/sys-api/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysApi'
              type: object
      security:
      - Bearer: []
      summary: 删除SysApi
      tags:
      - sys-SysApi
  /v2/admin/sys/sys-api/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysApi'
              type: object
      security:
      - Bearer: []
      summary: 获取SysApi
      tags:
      - sys-SysApi
  /v2/admin/sys/sys-api/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysApiGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.SysApi'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取SysApi列表
      tags:
      - sys-SysApi
  /v2/admin/sys/sys-api/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysApiDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysApi'
              type: object
      security:
      - Bearer: []
      summary: 更新SysApi
      tags:
      - sys-SysApi
  /v2/admin/sys/sys-menu-api/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysMenuApiDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysMenuApi'
              type: object
      security:
      - Bearer: []
      summary: 创建SysMenuApi
      tags:
      - sys-SysMenuApi
  /v2/admin/sys/sys-menu-api/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysMenuApi'
              type: object
      security:
      - Bearer: []
      summary: 删除SysMenuApi
      tags:
      - sys-SysMenuApi
  /v2/admin/sys/sys-menu-api/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysMenuApi'
              type: object
      security:
      - Bearer: []
      summary: 获取SysMenuApi
      tags:
      - sys-SysMenuApi
  /v2/admin/sys/sys-menu-api/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysMenuApiGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.SysMenuApi'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取SysMenuApi列表
      tags:
      - sys-SysMenuApi
  /v2/admin/sys/sys-menu-api/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: authorization
        type: string
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysMenuApiDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysMenuApi'
              type: object
      security:
      - Bearer: []
      summary: 更新SysMenuApi
      tags:
      - sys-SysMenuApi
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"